# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Blog API"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": " user login .",
                "operationId": "userlogin",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "user login",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    }
}